requirements.txt
celery==5.2.7
django-autocomplete-light==3.9.4
django-bootstrap3==22.1
django-debug-toolbar==3.7.0
django-dotenv==1.4.2
django-environ==0.9.0
django-jquery==3.1.0
django-querysetsequence==0.16
django-redis==5.2.0
django-schema-graph==2.2.1
django-select2==7.11.0
django-timezone-field==5.0
djangorestframework==3.14.0
environ==1.0
factory-boy==3.2.1
python-crontab==2.6.0
python-decouple==3.6
PyYAML==6.0
requests==2.28.1
typing==3.7.4.3
tzdata==2022.6
uritemplate==4.1.1
Django~=3.2.16